+++ compiler/symbol_table/scope.hpp
@@ -28,18 +28,18 @@ struct Scope {
     size_t starting_address;
     std::map<std::string, size_t> label_addresses;
     std::map<std::string, VarPtr> variables;
-    std::map<std::string, int> variable_indices;
+    std::map<std::string, size_t> variable_indices;
 	std::vector<std::string> args;
 	
-    int size = 0;
-    int temp_count = 0;
+    size_t size = 0;
+    size_t temp_count = 0;
 	mutable int label_count = 0;
 
     static std::shared_ptr<Scope> find_scope(const ScopePtr& scope, const std::string& name);
-	static int get_variable_index(const ScopePtr& scope, const std::string& name);
+	static size_t get_variable_index(const ScopePtr& scope, const std::string& name);
 
     Scope(SCOPE_TYPE type, const std::string& name) : name(name), type(type) {};
-    void generate_structure(int offset = 0);
+    void generate_structure(size_t offset = 0);
     bool has(std::string name);
     std::string get_full_name();
     std::string structure_to_string() const;
